solve task 3 and 4

+++ part3/hbnb/app/api/v1/reviews.py
@@ -1,6 +1,7 @@
  
 from flask_restx import Namespace, Resource, fields
 from app.services import facade
+from flask_jwt_extended import jwt_required, get_jwt_identity
 
 api = Namespace('reviews', description='Review operations')
 
@@ -19,9 +20,25 @@ class ReviewList(Resource):
     @api.expect(review_model)
     @api.response(201, 'Review successfully created')
     @api.response(400, 'Invalid input data')
+    @jwt_required()
     def post(self):
         """Register a new review"""
+        current_user_id = get_jwt_identity()
         review_data = api.payload
+
+        review_data['user_id'] = current_user_id
+        place_id = review_data['place_id']
+        place = facade.get_place(place_id)
+
+        if not place:
+            return {'error': 'place not found'}, 404
+        
+        if place.owner.id == current_user_id:
+            return {'error': 'You own the place you cant review'}, 400
+        
+        for review in place.reviews:
+            if review.user.id == current_user_id:
+                return {'error': 'You reviewd this place before'}, 400
         try:
             new_review = facade.create_review(review_data)
         except ValueError as e:
@@ -71,23 +88,33 @@ class ReviewResource(Resource):
     @api.response(200, 'Review updated successfully')
     @api.response(404, 'Review not found')
     @api.response(400, 'Invalid input data')
+    @jwt_required()
     def put(self, review_id):
         """Update a review's information"""
+        current_user_id = get_jwt_identity()
         review_data = api.payload
         review = facade.get_review(review_id)
         if not review:
             return {'error': 'Review not found'}, 404
+        
+        if review.user.id != current_user_id:
+            return {'error': 'Unauthrized action'}, 403
 
         facade.update_review(review_id, review_data)
         return {'message': 'Review updated successfully'}, 200
 
     @api.response(200, 'Review deleted successfully')
     @api.response(404, 'Review not found')
+    @jwt_required()
     def delete(self, review_id):
         """Delete a review"""
+        current_user_id = get_jwt_identity()
         review = facade.get_review(review_id)
         if not review:
             return {'error': 'Review not found'}, 404
+        
+        if review.user.id != current_user_id:
+            return {'error': 'Unauthorized action'}, 403
 
         facade.delete_review(review_id)
         return {'message': 'Review deleted successfully'}, 200
